Authorization, Policies and Gates

+++ app/Http/Controllers/PostController.php
@@ -7,6 +7,9 @@ use App\BlogPost;
 use Illuminate\Http\Request;
 use App\Http\Requests\StorePost;
 
+use Illuminate\Support\Facades\Gate;
+
+
 class PostController extends Controller
 {
     public function __construct(){
@@ -80,6 +83,13 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = BlogPost::findOrFail($id);
+
+        $this->authorize($post);
+
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You cannot edit another author's article");
+        // };
+
         return view('posts.edit', ['post' => $post]);
     }
 
@@ -93,6 +103,13 @@ class PostController extends Controller
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        $this->authorize($post);
+
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You cannot edit another author's article");
+        // };
+
         $validatedData = $request->validated();
 
         $post->fill($validatedData);
@@ -112,6 +129,14 @@ class PostController extends Controller
     public function destroy(Request $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        // if (Gate::denies('delete-post', $post)) {
+        //     abort(403, "You cannot delete another author's article");
+        // };
+
+        $this->authorize($post);
+
+
         $post->delete();
 
         // BlogPost::destroy($id);
